BNH-168 refactor: refactored tenant methods into tenant controller/service
