+++ web/Controllers/AdminController.cs
@@ -15,27 +15,18 @@ public class AdminController : AbstractController
     private readonly IAdminService _adminService;
     private readonly ILandlordService _landlordService;
     private readonly IPropertyService _propertyService;
-    private readonly ITenantService _tenantService;
-    private readonly ICsvImportService _csvImportService;
     private readonly ILogger<AdminController> _logger;
-    private readonly IMailService _mailService;
 
     public AdminController(
         ILogger<AdminController> logger,
         IAdminService adminService,
         ILandlordService landlordService,
-        IPropertyService propertyService,
-        ITenantService tenantService,
-        ICsvImportService csvImportService,
-        IMailService mailService)
+        IPropertyService propertyService)
     {
         _logger = logger;
         _adminService = adminService;
         _landlordService = landlordService;
         _propertyService = propertyService;
-        _tenantService = tenantService;
-        _csvImportService = csvImportService;
-        _mailService = mailService;
     }
 
     [HttpGet("/admin")]
@@ -147,72 +138,6 @@ public class AdminController : AbstractController
     }
 
     [Authorize(Roles = "Admin")]
-    [HttpGet]
-    public async Task<IActionResult> TenantList(HousingRequirementModel filter, string? targetPostcode, int page = 1,
-        int tenantsPerPage = 10)
-    {
-        if (targetPostcode != null)
-        {
-            var tenantsByLocation = await _tenantService.SortTenantsByLocation(targetPostcode, page, tenantsPerPage);
-
-            if (tenantsByLocation.Count > 0)
-            {
-                _logger.LogInformation("Successfully sorted by location");
-                return View(new TenantListModel(tenantsByLocation.TenantList, new HousingRequirementModel(),
-                    tenantsByLocation.Count, page, targetPostcode));
-            }
-
-            _logger.LogWarning($"Failed to find postcode {targetPostcode}");
-            AddFlashMessage("warning",
-                $"Failed to sort tenants using postcode {targetPostcode}: invalid postcode"); //TODO make this message actually display
-        }
-
-        var tenantsByFilter = await _adminService.GetTenantList(filter, page, tenantsPerPage);
-        return View(
-            new TenantListModel(tenantsByFilter.TenantList, filter, tenantsByFilter.Count, page, targetPostcode));
-    }
-
-    [Authorize(Roles = "Admin")]
-    [HttpGet("{currentPropertyId:int}/Matches")]
-    public async Task<IActionResult> TenantMatchList(int currentPropertyId)
-    {
-        var currentProperty =
-            PropertyViewModel.FromDbModel(_propertyService.GetPropertyByPropertyId(currentPropertyId)!);
-
-        var tenantMatchListModel = new TenantMatchListModel
-        {
-            TenantList = await _adminService.GetNearestTenantsToProperty(currentProperty),
-            CurrentProperty = currentProperty
-        };
-        return View("TenantMatchList", tenantMatchListModel);
-    }
-
-    [Authorize(Roles = "Admin")]
-    [HttpPost("{currentPropertyId:int}/Matches")]
-    public ActionResult SendMatchLinkEmail(string? propertyLink, string tenantEmail, int currentPropertyId)
-    {
-        if (propertyLink == null)
-        {
-            var property = PropertyViewModel.FromDbModel(_propertyService.GetPropertyByPropertyId(currentPropertyId)!);
-            if (property.PublicViewLink == null)
-            {
-                propertyLink = _propertyService.CreatePublicViewLink(currentPropertyId);
-            }
-        }
-
-        var addressToSendTo = new List<string> { tenantEmail };
-        var body = "Hi,\n\n" +
-                   "We have found a property that we think you might be interested in. Check it out at the link below:\n\n" +
-                   $"https://bricks-and-hearts.changeahead.org.uk/property/public/{propertyLink}\n\n" +
-                   "If you are interested, contact ChangeAhead to arrange a viewing.";
-
-        _mailService.TrySendMsgInBackground(body, "Check out this property from ChangeAhead", addressToSendTo);
-        _logger.LogInformation($"Successfully emailed tenant {tenantEmail}");
-        AddFlashMessage("success", $"Successfully emailed {tenantEmail}");
-        return RedirectToAction("TenantMatchList", new { currentPropertyId });
-    }
-
-    [Authorize(Roles = "Admin")]
     [HttpPost]
     public ActionResult GetInviteLink(int landlordId)
     {
@@ -284,56 +209,5 @@ public class AdminController : AbstractController
     public ActionResult GetSampleTenantCSV()
     {
         return File("~/TenantImportCSVTemplate.csv", "text/csv");
-    }
-
-    [Authorize(Roles = "Admin")]
-    [HttpPost]
-    public async Task<ActionResult> ImportTenants(IFormFile? csvFile)
-    {
-        if (csvFile == null)
-        {
-            return RedirectToAction(nameof(TenantList));
-        }
-
-        var fileName = csvFile.FileName;
-        if (csvFile.Length == 0)
-        {
-            _logger.LogWarning($"{fileName} has length zero.");
-            AddFlashMessage("danger",
-                $"{fileName} contains no data. Please upload a file containing the tenant data you would like to import.");
-            return RedirectToAction(nameof(TenantList));
-        }
-
-        if (fileName.Substring(fileName.Length - 3) != "csv")
-        {
-            _logger.LogWarning($"{fileName} not a CSV file.");
-            AddFlashMessage("danger", $"{fileName} is not a CSV file. Please upload your data as a CSV file.");
-            return RedirectToAction(nameof(TenantList));
-        }
-
-        var flashResponse = _csvImportService.CheckIfImportWorks(csvFile);
-        for (var i = 0; i < flashResponse.FlashMessages.Count; i++)
-        {
-            AddFlashMessage(flashResponse.FlashTypes[i], flashResponse.FlashMessages[i]);
-        }
-
-        if (flashResponse.FlashTypes.Contains("danger"))
-        {
-            return RedirectToAction(nameof(TenantList));
-        }
-
-        flashResponse = await _csvImportService.ImportTenants(csvFile);
-        for (var i = 0; i < flashResponse.FlashMessages.Count; i++)
-        {
-            AddFlashMessage(flashResponse.FlashTypes[i], flashResponse.FlashMessages[i]);
-        }
-
-        if (flashResponse.FlashMessages.Count == 0)
-        {
-            _logger.LogInformation("Successfuly imported all tenant data.");
-            AddFlashMessage("success", "Successfully imported all tenant data.");
-        }
-
-        return RedirectToAction(nameof(TenantList));
     }
 }
