+++ web/Views/Admin/TenantMatchList.cshtml
@@ -1,4 +1,4 @@
-﻿@model TenantMatchListModel
+@model TenantMatchListModel
 
 @{
     ViewData["Title"] = "Tenant List";
@@ -44,7 +44,7 @@
             <td>@tenant.HousingBenefits</td>
             <td>@tenant.Over35</td>
             <td>
-                @using (Html.BeginForm("SendMatchLinkEmail", "Admin", new
+                @using (Html.BeginForm("SendMatchLinkEmail", "Tenant", new
                 {
                     propertyLink = Model.CurrentProperty.PublicViewLink,
                     tenantEmail = tenant.Email,
