+++ web/Views/Property/ViewProperty.cshtml
@@ -34,7 +34,7 @@
 {
     <div class="row mt-4">
         <div class="col-auto mb-2">
-            <a asp-action="TenantMatchList" asp-controller="Admin" asp-route-currentPropertyId="@Model.Property.PropertyId" class="btn btn-large btn-outline-primary d-table-cell">View matches</a>
+            <a asp-action="TenantMatchList" asp-controller="Tenant" asp-route-currentPropertyId="@Model.Property.PropertyId" class="btn btn-large btn-outline-primary d-table-cell">View matches</a>
         </div>
         <div class="col-auto mb-2">
             <button role="button" id="copyBtn" class="btn btn-large btn-outline-primary" onclick="copyShareableLink('@Model.Property.PublicViewLink')">Copy shareable link</button>
