+++ web/Views/Shared/_Layout.cshtml
@@ -67,7 +67,7 @@
                 <div class="vr my-2"></div>
                 <a asp-controller="Admin" asp-action="LandlordList" class="nav-link text-light mx-3">Landlord list</a>
                 <div class="vr my-2"></div>
-                <a asp-controller="Admin" asp-action="TenantList" class="nav-link text-light mx-3">Tenant list</a>
+                <a asp-controller="Tenant" asp-action="TenantList" class="nav-link text-light mx-3">Tenant list</a>
                 <div class="vr my-2"></div>
                 <a asp-controller="Property" asp-action="PropertyList" class="nav-link text-light mx-3">Property list</a>
             </nav>
